requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
yfinance>=1.0.0
requests>=2.32.0
python-dotenv>=1.0.0
schedule>=1.2.2
pydantic>=2.0.0
beautifulsoup4>=4.12.0
numpy>=1.26.0
pandas>=2.0.0
